updated metadesc
