+++ public/index.html
@@ -5,7 +5,7 @@
         <link rel="icon" href="./pp.webp" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <meta name="theme-color" content="#000000" />
-        <meta name="description" content="dropsy" />
+        <meta name="description" content="a loot distribution tool for ff14" />
         <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
         <!-- *Fonts -->
         <link rel="preconnect" href="https://fonts.googleapis.com" />
