+++ src/components/bossKilled/bosses/P10s.js
