+++ src/components/characterSelect/characterCreate/CharacterCreate.js
@@ -3,11 +3,17 @@ import { Link } from "react-router-dom";
 
 const CharacterCreate = ({
     handleSubmit,
-    handleNameChange,
+    setCharName,
     gearPieces,
     setGearPieces,
 }) => {
     // *Letting useRadio work
+    // ? what does this mean
+
+    // *Handle Name Change
+    const handleNameChange = (e) => {
+        setCharName(e.target.value);
+    };
 
     return (
         <>
